+++ core/src/com/coronadefense/receiver/messages/PathToPathAnimationMessage.kt
@@ -1,10 +1,10 @@
 package com.coronadefense.receiver.messages
 
 data class PathToPathAnimationMessage(
-    val spriteNumber: UByte,
+    val spriteNumber: Int,
     val startPosition: Float,
     val endPosition: Float,
     val startTime: Float,
     val endTime: Float,
-    val resultAnimation: UByte,
+    val resultAnimation: Int,
 ): IMessage
