+++ core/src/com/coronadefense/receiver/messages/TowerAnimationMessage.kt
@@ -1,8 +1,8 @@
 package com.coronadefense.receiver.messages
 
 data class TowerAnimationMessage(
-    val towerId: UShort,
-    val animationNumber: UByte,
-    val rotation: Byte,
+    val towerId: Int,
+    val animationNumber: Int,
+    val rotation: Int,
     val time: Float,
 ): IMessage
